userService: update get, return null
