+++ package.json
@@ -19,9 +19,9 @@
     "@types/express-fileupload": "^1.1.6",
     "@types/express-session": "^1.17.3",
     "@types/jest": "^26.0.22",
+    "@types/lodash.throttle": "^4.1.6",
     "@types/passport": "^1.0.6",
     "@types/passport-local": "^1.0.33",
-    "@types/lodash.throttle": "^4.1.6",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
     "copyfiles": "^2.4.1",
@@ -39,7 +39,7 @@
   },
   "dependencies": {
     "@godaddy/terminus": "^4.7.1",
-    "@openforis/arena-core": "^0.0.59",
+    "@openforis/arena-core": "^0.0.62",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "compression": "^1.7.4",
