+++ src/server/middleware/authentication.ts
@@ -1,54 +1,10 @@
 import passport from 'passport'
 import { VerifyFunctionWithRequest, Strategy as LocalStrategy } from 'passport-local'
-import bcrypt from 'bcryptjs'
 import { Express } from 'express'
-import { FieldValidators, User, UserStatus, Validator } from '@openforis/arena-core'
+import { FieldValidators, User, UserStatus, Validator, ValidatorErrorKeys } from '@openforis/arena-core'
 
 import { ExpressInitializer } from '../expressInitializer'
 import { UserServiceServer } from '../../service'
-import { UserResetPasswordRepository, AuthGroupRepository } from '../../repository'
-
-const comparePassword = bcrypt.compareSync
-
-// TODO: Handle these separately
-const ValidationMessages = {
-  emailInvalid: 'validationErrors.user.emailInvalid',
-  userNotFound: 'validationErrors.user.userNotFound',
-  passwordChangeRequired: 'validationErrors.user.passwordChangeRequired',
-}
-
-const _initializeUser = async (_user: User): Promise<User> => {
-  const user = { ..._user }
-  if (user.password) delete user.password
-
-  // Assoc auth groups
-  let userUpdated = {
-    ...user,
-    authGroups: await AuthGroupRepository.fetchUserGroups({ userUuid: user.uuid }),
-  }
-  if (user.status === UserStatus.INVITED) {
-    const expired = !(await UserResetPasswordRepository.hasValidResetPassword({ userUuid: user.uuid }))
-    userUpdated = {
-      ...userUpdated,
-      invitation: {
-        expired,
-      },
-    }
-  }
-
-  // @ts-ignore fixed by implementing repository method
-  return userUpdated
-}
-
-const findUserByEmailAndPassword = async (email: string, password: string) => {
-  const user = await UserServiceServer.get({ email })
-
-  if (user?.password && (await comparePassword(password, user.password))) {
-    return _initializeUser(user)
-  }
-
-  return null
-}
 
 const _verifyCallback: VerifyFunctionWithRequest = async (_, email, password, done) => {
   const sendUser = (user: User) => done(null, user)
@@ -62,14 +18,14 @@ const _verifyCallback: VerifyFunctionWithRequest = async (_, email, password, do
   )
 
   if (!isValidUser) {
-    sendError(ValidationMessages.emailInvalid)
+    sendError(ValidatorErrorKeys.user.emailInvalid)
     return
   }
 
-  const user = await findUserByEmailAndPassword(email, password)
+  const user = await UserServiceServer.get({ email, password })
 
   if (!user) {
-    sendError(ValidationMessages.userNotFound)
+    sendError(ValidatorErrorKeys.user.userNotFound)
     return
   }
 
@@ -77,7 +33,7 @@ const _verifyCallback: VerifyFunctionWithRequest = async (_, email, password, do
     sendUser(user)
   }
 
-  sendError(ValidationMessages.passwordChangeRequired)
+  sendError(ValidatorErrorKeys.user.passwordChangeRequired)
 }
 
 const localStrategy = new LocalStrategy(
