+++ src/utils/requests.ts
@@ -1,4 +1,5 @@
 import { Request } from 'express'
+import { User } from '@openforis/arena-core'
 
 const getHost = (request: Request): string | undefined => request.header('host')
 const getParams = (req: Request): Record<string, any> => {
@@ -12,7 +13,7 @@ const getParams = (req: Request): Record<string, any> => {
 
 const getServerUrl = (request: Request): string => `${request.protocol}://${request.get('host')}`
 const getUrl = (request: Request): string => request.url
-const getUser = (request: Request) => request.user
+const getUser = (request: Request): User => request.user
 const isHttps = (request: Request): boolean => request.secure || request.header('x-forwarded-proto') === 'https'
 
 export const Requests = {
